+++ coxeter/__init__.py
@@ -4,6 +4,7 @@ from . import symmetry
 from . import ft
 from . import shape_classes
 from . import common_shapes
+from .shape_getters import from_gsd_type_shapes
 
 __all__ = [
     'shapes',
@@ -12,6 +13,7 @@ __all__ = [
     'ft',
     'shape_classes',
     'common_shapes',
+    'from_gsd_type_shapes'
 ]
 
 __version__ = '0.2.0'