+++ coxeter/shape_classes/convex_spheropolyhedron.py
@@ -8,7 +8,7 @@ class ConvexSpheropolyhedron(object):
 
         A convex spheropolyhedron is defined as a convex polyhedron plus a
         rounding radius. All properties of the underlying polyhedron (the
-        vertices, the facets and their neighbors, etc) can be accessed directly
+        vertices, the faces and their neighbors, etc) can be accessed directly
         through :attr:`.polyhedron`.
 
         Args:
@@ -21,10 +21,24 @@ class ConvexSpheropolyhedron(object):
         self._radius = radius
 
     @property
+    def gsd_shape_spec(self):
+        """dict: A complete description of this shape corresponding to the
+        shape specification in the GSD file format as described
+        `here <https://gsd.readthedocs.io/en/stable/shapes.html>`_."""
+        return {'type': 'ConvexPolyhedron',
+                'vertices': self.polyhedron._vertices.tolist(),
+                'rounding_radius': self._radius}
+
+    @property
     def polyhedron(self):
         """:class:`~coxeter.shape_classes.convex_polyhedron.ConvexPolyhedron`:
         The underlying polyhedron."""
         return self._polyhedron
+
+    @property
+    def vertices(self):
+        """Get the vertices of the spheropolyhedron."""
+        return self._polyhedron.vertices
 
     @property
     def volume(self):
@@ -36,7 +50,7 @@ class ConvexSpheropolyhedron(object):
         # For every pair of faces, find the dihedral angle, divide by 2*pi to
         # get the fraction of a cylinder it includes, then multiply by the edge
         # length to get the cylinder contribution.
-        for i, j, edge in self.polyhedron._get_facet_intersections():
+        for i, j, edge in self.polyhedron._get_face_intersections():
             phi = self.polyhedron.get_dihedral(i, j)
             edge_length = np.linalg.norm(self.polyhedron.vertices[edge[0]] -
                                          self.polyhedron.vertices[edge[1]])
@@ -59,7 +73,7 @@ class ConvexSpheropolyhedron(object):
         # For every pair of faces, find the dihedral angle, divide by 2*pi to
         # get the fraction of a cylinder it includes, then multiply by the edge
         # length to get the cylinder contribution.
-        for i, j, edge in self.polyhedron._get_facet_intersections():
+        for i, j, edge in self.polyhedron._get_face_intersections():
             phi = self.polyhedron.get_dihedral(i, j)
             edge_length = np.linalg.norm(self.polyhedron.vertices[edge[0]] -
                                          self.polyhedron.vertices[edge[1]])
@@ -85,7 +99,7 @@ class ConvexSpheropolyhedron(object):
                 spheropolyhedron.
         """  # noqa: E501
         # Determine which points are in the polyhedron and which are in the
-        # bounded volume of facets extruded by the rounding radius
+        # bounded volume of faces extruded by the rounding radius
         points = np.atleast_2d(points)
         point_plane_distances = self.polyhedron._point_plane_distances(points)
         in_polyhedron = np.all(point_plane_distances <= 0, axis=1)
@@ -94,62 +108,62 @@ class ConvexSpheropolyhedron(object):
         if np.all(in_polyhedron):
             return in_polyhedron
 
-        # Compute extrusions of the facets
-        extruded_facets = []
-        for facet, normal in zip(self.polyhedron.facets,
-                                 self.polyhedron.normals):
-            base_vertices = self.polyhedron.vertices[facet]
+        # Compute extrusions of the faces
+        extruded_faces = []
+        for face, normal in zip(self.polyhedron.faces,
+                                self.polyhedron.normals):
+            base_vertices = self.polyhedron.vertices[face]
             extruded_vertices = base_vertices + self.radius * normal
-            extruded_facets.append(
+            extruded_faces.append(
                 ConvexPolyhedron([*base_vertices, *extruded_vertices]))
 
         # Select the points between the inner polyhedron and extruded space
-        # and then filter them using the point-facet distances
-        point_facets_in_polyhedron_hull = point_plane_distances <= 0
-        point_facets_in_extruded_hull = point_plane_distances <= self.radius
-        point_facets_to_check = \
-            point_facets_in_extruded_hull & ~point_facets_in_polyhedron_hull
+        # and then filter them using the point-face distances
+        point_faces_in_polyhedron_hull = point_plane_distances <= 0
+        point_faces_in_extruded_hull = point_plane_distances <= self.radius
+        point_faces_to_check = \
+            point_faces_in_extruded_hull & ~point_faces_in_polyhedron_hull
 
         # Exit early if there are no intersections to check between points
-        # and rounded facets
-        if not np.any(point_facets_to_check):
+        # and rounded faces
+        if not np.any(point_faces_to_check):
             return in_polyhedron
 
-        def check_facet(point_id, facet_id):
-            """Checks for intersection of the point with rounded facets."""
+        def check_face(point_id, face_id):
+            """Checks for intersection of the point with rounded faces."""
             point = points[point_id]
-            extruded_facet = extruded_facets[facet_id]
-            in_extruded_facet = extruded_facet.is_inside(point)[0]
+            extruded_face = extruded_faces[face_id]
+            in_extruded_face = extruded_face.is_inside(point)[0]
 
-            # Exit early if the point is found in the extruded facet
-            if in_extruded_facet:
+            # Exit early if the point is found in the extruded face
+            if in_extruded_face:
                 return True
 
             # Check spherocylinders around the edges (excluding spherical caps)
-            facet_points = self.polyhedron.vertices[
-                self.polyhedron.facets[facet_id]]
+            face_points = self.polyhedron.vertices[
+                self.polyhedron.faces[face_id]]
 
-            # Vectors along the facet edges
-            facet_edges = np.roll(facet_points, -1, axis=0) - facet_points
+            # Vectors along the face edges
+            face_edges = np.roll(face_points, -1, axis=0) - face_points
 
-            # Normalized vectors along the facet edges
-            facet_edge_lengths = np.linalg.norm(facet_edges, axis=-1)
-            facet_edges_norm = facet_edges / facet_edge_lengths[:, np.newaxis]
+            # Normalized vectors along the face edges
+            face_edge_lengths = np.linalg.norm(face_edges, axis=-1)
+            face_edges_norm = face_edges / face_edge_lengths[:, np.newaxis]
 
             # Vectors from point to the edge starts
-            point_to_edge_starts = point - facet_points
+            point_to_edge_starts = point - face_points
 
             # Compute the vector rejection (perpendicular projection) of point
             # along edge vectors and determine if the cylinders contain it
-            edge_projections = np.sum(point_to_edge_starts * facet_edges_norm,
+            edge_projections = np.sum(point_to_edge_starts * face_edges_norm,
                                       axis=1)
             perpendicular_projections = point_to_edge_starts - \
-                edge_projections[:, np.newaxis] * facet_edges_norm
+                edge_projections[:, np.newaxis] * face_edges_norm
             cylinder_distances = np.linalg.norm(
                 perpendicular_projections, axis=-1)
             in_cylinders = np.any((cylinder_distances <= self.radius) &
                                   (edge_projections >= 0) &
-                                  (edge_projections <= facet_edge_lengths))
+                                  (edge_projections <= face_edge_lengths))
 
             # Exit early if the point is found in the cylinders
             if in_cylinders:
@@ -160,10 +174,10 @@ class ConvexSpheropolyhedron(object):
             in_caps = np.any(cap_distances <= self.radius)
             return in_caps
 
-        # Check the facets whose rounded portion could contain the point
+        # Check the faces whose rounded portion could contain the point
         in_sphero_shape = np.zeros(len(points), dtype=bool)
-        for point_id, facet_id in zip(*np.where(point_facets_to_check)):
+        for point_id, face_id in zip(*np.where(point_faces_to_check)):
             if not in_sphero_shape[point_id]:
-                in_sphero_shape[point_id] = check_facet(point_id, facet_id)
+                in_sphero_shape[point_id] = check_face(point_id, face_id)
 
         return in_polyhedron | in_sphero_shape