+++ coxeter/shape_classes/ellipse.py
@@ -21,6 +21,13 @@ class Ellipse(Shape2D):
         self._center = np.asarray(center)
 
     @property
+    def gsd_shape_spec(self):
+        """dict: A complete description of this shape corresponding to the
+        shape specification in the GSD file format as described
+        `here <https://gsd.readthedocs.io/en/stable/shapes.html>`_."""
+        return {'type': 'Ellipsoid', 'a': self._a, 'b': self._b}
+
+    @property
     def center(self):
         return self._center